pyproject.toml
[project]
name = "ai-repurposer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-google-genai>=4.2.0",
    "langgraph>=1.0.7",
    "moviepy>=2.2.1",
    "python-dotenv>=1.2.1",
    "streamlit>=1.52.2",
    "youtube-transcript-api>=1.2.4",
    "yt-dlp>=2026.1.29",
]
